requirements.txt
requests
gensim
python-docx
unidecode
nltk
numpy
matplotlib
genson
flatdict
pandas
sh
pytest
pylint
osfclient
urllib3
rich
pyyaml
mock